+++ pilz_control/CMakeLists.txt
@@ -145,10 +145,17 @@ if(CATKIN_ENABLE_TESTING)
   add_rostest_gtest(unittest_pilz_joint_trajectory_controller
     test/unittest_pilz_joint_trajectory_controller.test
     test/unittest_pilz_joint_trajectory_controller.cpp
+    test/unittest_pilz_joint_trajectory_controller_is_executing.cpp
     test/robot_mock.cpp
     src/cartesian_speed_monitor.cpp
   )
   target_link_libraries(unittest_pilz_joint_trajectory_controller ${catkin_LIBRARIES})
+
+  add_rostest_gtest(unittest_get_joint_acceleration_limits
+    test/unittest_get_joint_acceleration_limits.test
+    test/unittest_get_joint_acceleration_limits.cpp
+  )
+  target_link_libraries(unittest_get_joint_acceleration_limits ${catkin_LIBRARIES})
 
   add_rostest(test/integrationtest_pilz_joint_trajectory_controller.test
     DEPENDENCIES ${PROJECT_NAME} robot_mock
