+++ pilz_control/test/unittest_pilz_joint_trajectory_controller.cpp
@@ -14,22 +14,17 @@
  * limitations under the License.
  */
 
-#include <chrono>
-#include <functional>
-#include <future>
+#include <memory>
+#include <string>
 #include <thread>
+#include <vector>
 
 #include <gtest/gtest.h>
 
 #include <ros/ros.h>
-#include <joint_trajectory_controller/hardware_interface_adapter.h>
 
-#include <control_msgs/FollowJointTrajectoryAction.h>
-#include <trajectory_msgs/JointTrajectory.h>
 #include <std_srvs/Trigger.h>
 #include <hardware_interface/joint_command_interface.h>
-#include <hardware_interface/joint_state_interface.h>
-#include <hardware_interface/robot_hw.h>
 #include <trajectory_interface/quintic_spline_segment.h>
 
 #include <pilz_control/pilz_joint_trajectory_controller.h>
@@ -47,12 +42,6 @@ static const std::string TRAJECTORY_ACTION{ "/follow_joint_trajectory" };
 static const std::string HOLD_SERVICE{ "/hold" };
 static const std::string UNHOLD_SERVICE{ "/unhold" };
 static const std::string IS_EXECUTING_SERVICE{ "/is_executing" };
-static const std::string TRAJECTORY_COMMAND_TOPIC{ "/command" };
-static const std::string CONTROLLER_JOINT_NAMES_PARAM{ "/controller_joint_names" };
-
-static const std::string JOINT_WITH_HAS_ACC_LIM_FALSE{ "joint_with_has_acc_lim_false" };
-static const std::string JOINT_WITH_UNDEFINED_MAX_ACC{ "joint_with_undefined_max_acc" };
-static const std::string JOINT_WITH_UNDEFINED_HAS_ACC_LIM{ "joint_with_undefined_has_acc_lim" };
 
 using namespace pilz_joint_trajectory_controller;
 
@@ -62,32 +51,6 @@ using PJTCManager = PJTCManagerMock<Segment, HWInterface>;
 using RobotDriver = RobotDriverMock<PJTCManager>;
 using Controller = pilz_joint_trajectory_controller::PilzJointTrajectoryController<Segment, HWInterface>;
 using ControllerPtr = std::shared_ptr<Controller>;
-
-/**
- * @brief Dummy needed for testGetJointAccelerationLimits.
- */
-class DummySegmentImpl
-{
-public:
-  struct State : public Segment::State
-  {
-  };
-  typedef typename Segment::Scalar Scalar;
-  typedef typename Segment::Time Time;
-};
-
-/**
- * @brief Dummy needed for testGetJointAccelerationLimits.
- */
-class DummyHardwareInterface : public hardware_interface::RobotHW
-{
-public:
-  struct MyResourceHandleType
-  {
-  };
-  typedef typename DummyHardwareInterface::MyResourceHandleType ResourceHandleType;
-  typedef typename hardware_interface::JointHandle JointHandle;
-};
 
 /**
  * @brief Test fixture class for the unit-test of the PilzJointTrajectoryController.
@@ -103,22 +66,18 @@ protected:
 
   testing::AssertionResult isControllerInHoldMode();
   testing::AssertionResult isControllerInUnholdMode();
-  void runGetJointAccelerationLimits(const ros::NodeHandle& nh, const std::vector<std::string>& joint_names);
 
 protected:
   RobotDriver robot_driver_{ CONTROLLER_NAMESPACE };
   std::shared_ptr<PJTCManager> manager_;
   TrajectoryActionClientWrapper action_client_{ CONTROLLER_NAMESPACE + TRAJECTORY_ACTION };
-  ros::NodeHandle controller_nh_{ CONTROLLER_NAMESPACE };
   ros::AsyncSpinner spinner_{ 2 };
-  std::vector<std::string> controller_joint_names_;
 };
 
 void PilzJointTrajectoryControllerTest::SetUp()
 {
   spinner_.start();
   manager_ = robot_driver_.getManager();
-  controller_joint_names_ = std::vector<std::string>(JOINT_NAMES.begin(), JOINT_NAMES.end());
   setControllerParameters(CONTROLLER_NAMESPACE);
   startSimTime();
 }
@@ -155,12 +114,6 @@ testing::AssertionResult PilzJointTrajectoryControllerTest::isControllerInUnhold
   return testing::AssertionSuccess();
 }
 
-void PilzJointTrajectoryControllerTest::runGetJointAccelerationLimits(const ros::NodeHandle& nh,
-                                                                      const std::vector<std::string>& joint_names)
-{
-  PilzJointTrajectoryController<DummySegmentImpl, DummyHardwareInterface>::getJointAccelerationLimits(nh, joint_names);
-}
-
 //////////////////////////////////
 //  Testing of Initialization   //
 //////////////////////////////////
@@ -176,10 +129,6 @@ TEST_F(PilzJointTrajectoryControllerTest, testInitializiation)
   EXPECT_TRUE(action_client_.waitForActionServer());
 }
 
-/////////////////////////////////////////
-//    Testing of additional methods    //
-/////////////////////////////////////////
-
 /**
  * @brief Test increases function coverage by ensuring that all Dtor variants are called.
  */
@@ -187,71 +136,6 @@ TEST_F(PilzJointTrajectoryControllerTest, testD0Destructor)
 {
   ControllerPtr controller{ new Controller() };
   SUCCEED();
-}
-
-/**
- * @tests{Monitor_joint_accelerations,
- *   Test if parameters for acceleration limits are correctly read.
- * }
- */
-TEST_F(PilzJointTrajectoryControllerTest, testGetJointAccelerationLimits)
-{
-  ros::NodeHandle limits_nh(controller_nh_, LIMITS_NAMESPACE);
-  const std::vector<boost::optional<double>> acceleration_limits =
-      PilzJointTrajectoryController<DummySegmentImpl, DummyHardwareInterface>::getJointAccelerationLimits(
-          limits_nh, controller_joint_names_);
-  EXPECT_EQ(controller_joint_names_.size(), acceleration_limits.size());
-  EXPECT_TRUE(acceleration_limits.at(0));
-  EXPECT_TRUE(acceleration_limits.at(1));
-  EXPECT_DOUBLE_EQ(*acceleration_limits.at(0), MAX_JOINT_ACCELERATION);
-  EXPECT_DOUBLE_EQ(*acceleration_limits.at(1), MAX_JOINT_ACCELERATION);
-}
-
-/**
- * @tests{Monitor_joint_accelerations,
- *   Test if parameters for acceleration limits are omitted for joints which have no limit.
- * }
- */
-TEST_F(PilzJointTrajectoryControllerTest, testGetJointAccelerationLimitsHasNoLimit)
-{
-  ros::NodeHandle nh{ "~" };
-  std::stringstream full_param_name;
-  full_param_name << JOINT_WITH_HAS_ACC_LIM_FALSE << "/" << HAS_ACCELERATION_PARAMETER;
-  nh.setParam(full_param_name.str(), false);
-
-  const std::vector<boost::optional<double>> acceleration_limits_has_acc_lim_false =
-      PilzJointTrajectoryController<DummySegmentImpl, DummyHardwareInterface>::getJointAccelerationLimits(
-          nh, { JOINT_WITH_HAS_ACC_LIM_FALSE });
-  EXPECT_EQ(1U, acceleration_limits_has_acc_lim_false.size());
-  EXPECT_FALSE(acceleration_limits_has_acc_lim_false.at(0));
-}
-
-/**
- * @tests{Monitor_joint_accelerations,
- *   Test if correct errors are thrown if parameters can not be correctly read.
- * }
- */
-TEST_F(PilzJointTrajectoryControllerTest, testGetJointAccelerationLimitUndefinedLimit)
-{
-  ros::NodeHandle nh{ "~" };
-  std::stringstream full_param_name;
-  full_param_name << JOINT_WITH_UNDEFINED_MAX_ACC << "/" << HAS_ACCELERATION_PARAMETER;
-  nh.setParam(full_param_name.str(), true);
-
-  EXPECT_THROW(runGetJointAccelerationLimits(nh, { JOINT_WITH_UNDEFINED_MAX_ACC }), ros::InvalidParameterException);
-}
-
-/**
- * @tests{Monitor_joint_accelerations,
- *   Test if correct errors are thrown if parameters can not be correctly read.
- * }
- */
-TEST_F(PilzJointTrajectoryControllerTest, testGetJointAccelerationLimitUndefinedFlag)
-{
-  ros::NodeHandle nh{ "~" };
-  nh.setParam(JOINT_WITH_UNDEFINED_HAS_ACC_LIM, true);
-
-  EXPECT_THROW(runGetJointAccelerationLimits(nh, { JOINT_WITH_UNDEFINED_HAS_ACC_LIM }), ros::InvalidParameterException);
 }
 
 /////////////////////////////////////
@@ -513,195 +397,6 @@ TEST_F(PilzJointTrajectoryControllerTest, testTrajectoryWithTooHighAcceleration)
         << "Difference in position despite no movement should have been performed.";
   }
 }
-
-//////////////////////////////////////////////////////////////////////////
-//    Parameterized tests for the "is-executing-check" functionality    //
-//////////////////////////////////////////////////////////////////////////
-
-//! The return value indicates if the call was successful (in case of a service callback), the actual result
-//! of the is-executing-check is assigned (via reference) to the second argument.
-using InvokeIsExecuting = std::function<testing::AssertionResult(const ControllerPtr&, bool&)>;
-
-static testing::AssertionResult InvokeIsExecutingMethod(const ControllerPtr& controller, bool& result)
-{
-  result = controller->is_executing();
-  return testing::AssertionSuccess();
-}
-
-static testing::AssertionResult InvokeIsExecutingServiceCallback(const ControllerPtr& controller, bool& result)
-{
-  std_srvs::TriggerRequest req;
-  std_srvs::TriggerResponse resp;
-  if (!controller->handleIsExecutingRequest(req, resp))
-  {
-    return testing::AssertionFailure() << "Callback of is_executing returned false unexpectedly.";
-  }
-
-  result = resp.success;
-  return testing::AssertionSuccess();
-}
-
-/**
- * @brief For testing both the isExecuting method and the is_executing service callback we use parameterized tests.
- */
-class PilzJointTrajectoryControllerIsExecutingTest : public testing::Test,
-                                                     public testing::WithParamInterface<InvokeIsExecuting>
-{
-protected:
-  void SetUp() override;
-
-  testing::AssertionResult invokeIsExecuting(bool& result);
-
-protected:
-  RobotDriver robot_driver_{ CONTROLLER_NAMESPACE };
-  std::shared_ptr<PJTCManager> manager_;
-  TrajectoryActionClientWrapper action_client_{ CONTROLLER_NAMESPACE + TRAJECTORY_ACTION };
-  ros::NodeHandle controller_nh_{ CONTROLLER_NAMESPACE };
-  ros::AsyncSpinner spinner_{ 2 };
-};
-
-void PilzJointTrajectoryControllerIsExecutingTest::SetUp()
-{
-  spinner_.start();
-  manager_ = robot_driver_.getManager();
-  setControllerParameters(CONTROLLER_NAMESPACE);
-  startSimTime();
-}
-
-testing::AssertionResult PilzJointTrajectoryControllerIsExecutingTest::invokeIsExecuting(bool& result)
-{
-  auto invoke_is_executing{ GetParam() };
-  return invoke_is_executing(manager_->controller_, result);
-}
-
-TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testNotStarted)
-{
-  ASSERT_TRUE(manager_->loadController()) << "Failed to initialize the controller.";
-
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
-}
-
-/**
- * @brief Check that a fake start (simply setting state to RUNNING) doesn't trigger that is_executing() returns true.
- *
- * Increases line coverage.
- */
-TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testFakeStart)
-{
-  ASSERT_TRUE(manager_->loadController()) << "Failed to initialize the controller.";
-
-  manager_->controller_->state_ = Controller::RUNNING;
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
-}
-
-TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testStopTrajExecutionAtStart)
-{
-  ASSERT_TRUE(manager_->loadController()) << "Failed to initialize the controller.";
-
-  manager_->startController();
-
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_TRUE(is_executing_result) << "Failed to detect stop trajectory execution at start.";
-
-  ros::Duration stop_duration{ STOP_TRAJECTORY_DURATION_SEC + 2 * DEFAULT_UPDATE_PERIOD_SEC };
-  progressInTime(stop_duration);
-  robot_driver_.update();
-
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
-}
-
-TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testNotExecutingAfterUnhold)
-{
-  ASSERT_TRUE(performFullControllerStartup(&robot_driver_));
-
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
-}
-
-TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testActionGoalExecution)
-{
-  ASSERT_TRUE(performFullControllerStartup(&robot_driver_));
-
-  GoalType goal{ generateAlternatingGoal<RobotDriver>(&robot_driver_) };
-  action_client_.sendGoal(goal);
-
-  EXPECT_TRUE(updateUntilRobotMotion<RobotDriver>(&robot_driver_));
-
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_TRUE(is_executing_result) << "Failed to detect action goal execution";
-
-  progressInTime(getGoalDuration(goal) + ros::Duration(DEFAULT_UPDATE_PERIOD_SEC));
-  robot_driver_.update();
-
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
-}
-
-TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testTrajCommandExecution)
-{
-  ASSERT_TRUE(performFullControllerStartup(&robot_driver_));
-
-  ros::NodeHandle nh{ "~" };
-  ros::Publisher trajectory_command_publisher =
-      nh.advertise<trajectory_msgs::JointTrajectory>(CONTROLLER_NAMESPACE + TRAJECTORY_COMMAND_TOPIC, 1);
-
-  GoalType goal{ generateAlternatingGoal<RobotDriver>(&robot_driver_) };
-  trajectory_command_publisher.publish(goal.trajectory);
-
-  EXPECT_TRUE(updateUntilRobotMotion<RobotDriver>(&robot_driver_));
-
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_TRUE(is_executing_result) << "Failed to detect trajectory command execution";
-
-  progressInTime(getGoalDuration(goal) + ros::Duration(DEFAULT_UPDATE_PERIOD_SEC));
-  robot_driver_.update();
-
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
-}
-
-TEST_P(PilzJointTrajectoryControllerIsExecutingTest, testStopTrajExecutionAtHold)
-{
-  ASSERT_TRUE(performFullControllerStartup(&robot_driver_));
-
-  GoalType goal{ generateAlternatingGoal<RobotDriver>(&robot_driver_) };
-  action_client_.sendGoal(goal);
-
-  EXPECT_TRUE(updateUntilRobotMotion<RobotDriver>(&robot_driver_));
-
-  std_srvs::TriggerRequest req;
-  std_srvs::TriggerResponse resp;
-  // run async such that stop trajectory can be executed in the meantime
-  std::future<bool> hold_future = manager_->triggerHoldAsync(req, resp);
-
-  EXPECT_TRUE(action_client_.waitForActionResult());
-  // the following fails due to https://github.com/ros-controls/ros_controllers/issues/174
-  // EXPECT_EQ(action_client_.getResult()->error_code, control_msgs::FollowJointTrajectoryResult::INVALID_GOAL);
-
-  robot_driver_.update();
-
-  bool is_executing_result;
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_TRUE(is_executing_result) << "Failed to detect stop trajectory execution";
-
-  EXPECT_TRUE(updateUntilHoldMode<RobotDriver>(&robot_driver_, hold_future));
-  EXPECT_TRUE(resp.success);
-
-  EXPECT_TRUE(invokeIsExecuting(is_executing_result));
-  EXPECT_FALSE(is_executing_result) << "There should be no execution to detect";
-}
-
-INSTANTIATE_TEST_CASE_P(MethodAndServiceCallback, PilzJointTrajectoryControllerIsExecutingTest,
-                        testing::Values(InvokeIsExecutingMethod, InvokeIsExecutingServiceCallback));
 
 }  // namespace pilz_joint_trajectory_controller_test
